refactor: refactor state-manager
